feat: sketch out micropipeline
We don't need analyis code as complex as it is in Web Connectivity
but it would still help to reconstruct what happened by sorting
such that we can understand which targets failed.

This commit is an initial stepping stone for implementing this
functionality. What remains to be done is to have a way to sort
linearized observations according to specific set of keys.