feat(signal): fully implement richer input
It's possible to run in richer input mode only when the user
has used the `-m` flag to run a module. In such a case, using
also the `-f` flag allows to specify a file w/ richer input.

The format of the richer input is very custom to what we need
to currently do for signal. This code is obviously experimental
but sheds light on how richer input could be implemented.